+++ src/Lightbox.js
@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { css, StyleSheet } from 'aphrodite/no-important';
+import { css, StyleSheet } from 'aphrodite';
+import ScrollLock from 'react-scrolllock';
+import * as ReactSpinners from 'react-spinners';
 
-import theme from './theme';
+import defaultTheme from './theme';
 import Arrow from './components/Arrow';
 import Container from './components/Container';
 import Footer from './components/Footer';
@@ -10,45 +12,50 @@ import Header from './components/Header';
 import PaginatedThumbnails from './components/PaginatedThumbnails';
 import Portal from './components/Portal';
 
-import { bindFunctions, canUseDom } from './utils';
+import bindFunctions from './utils/bindFunctions';
+import canUseDom from './utils/canUseDom';
+import deepMerge from './utils/deepMerge';
 
 class Lightbox extends Component {
-	constructor () {
-		super();
-		this.state = { rotate: 0, isZoomed: false };
+	constructor (props) {
+		super(props);
+
+		this.theme = deepMerge(defaultTheme, props.theme);
+		this.classes = StyleSheet.create(deepMerge(defaultStyles, this.theme));
+		this.state = { imageLoaded: false, rotate: 0, isZoomed: false };
 
 		bindFunctions.call(this, [
 			'gotoNext',
 			'gotoPrev',
 			'rotate',
 			'zoom',
+			'closeBackdrop',
 			'handleKeyboardInput',
+			'handleImageLoaded',
 		]);
 	}
 	getChildContext () {
 		return {
-			theme: this.props.theme,
+			theme: this.theme,
 		};
 	}
 	componentDidMount () {
-		let self = this;
-
 		if (this.props.isOpen && this.props.enableKeyboardInput) {
 			window.addEventListener('keydown', this.handleKeyboardInput);
 		}
 
-		if (this.props.zoom) {
-			document.addEventListener('mousemove', function (event) {
-				let posY = event.clientY;
-				if (self.state.isZoomed) {
-					if (posY <= window.innerHeight / 2) {
-						self.setState({ margin: `${window.innerHeight - posY}px 0px 0px 0px` });
-					} else {
-						self.setState({ margin: `-${posY / 1.3}px 0px 0px 0px` });
-					}
-				}
-			});
-		}
+        if (this.props.zoom) {
+            document.addEventListener('mousemove', (event) => {
+                let posY = event.clientY;
+                if (this.state.isZoomed) {
+                    if (posY <= window.innerHeight / 2) {
+                    	this.setState({ margin: `${window.innerHeight - posY}px 0 0 0` });
+                    } else {
+                    	this.setState({ margin: `-${posY / 1.3}px 0 0 0` });
+                    }
+                }
+            });
+        }
 
 	}
 	componentWillReceiveProps (nextProps) {
@@ -77,6 +84,12 @@ class Lightbox extends Component {
 			}
 		}
 
+		// preload current image
+		if (this.props.currentImage !== nextProps.currentImage || !this.props.isOpen && nextProps.isOpen) {
+			const img = this.preloadImage(nextProps.currentImage, this.handleImageLoaded);
+			this.setState({ imageLoaded: img.complete });
+		}
+
 		// add/remove event listeners
 		if (!this.props.isOpen && nextProps.isOpen && nextProps.enableKeyboardInput) {
 			window.addEventListener('keydown', this.handleKeyboardInput);
@@ -95,39 +108,62 @@ class Lightbox extends Component {
 	// METHODS
 	// ==============================
 
-	preloadImage (idx) {
+	preloadImage (idx, onload) {
 		const image = this.props.images[idx];
-
 		if (!image) return;
 
 		const img = new Image();
 
+		// TODO: add error handling for missing images
+		img.onerror = onload;
+		img.onload = onload;
 		img.src = image.src;
+		img.srcSet = image.srcSet || image.srcset;
 
 		if (image.srcset) {
 			img.srcset = image.srcset.join();
 		}
 
+		if (img.srcSet) img.setAttribute('srcset', img.srcSet);
+
+		return img;
 	}
 	gotoNext (event) {
-		if (this.props.currentImage === (this.props.images.length - 1)) return;
+		const { currentImage, images } = this.props;
+		const { imageLoaded } = this.state;
+
+		if (!imageLoaded || currentImage === (images.length - 1)) return;
+
 		if (event) {
 			event.preventDefault();
 			event.stopPropagation();
 		}
+
 		this.props.onClickNext();
 		this.setState({ rotate: 0 });
 
+		this.props.onClickNext();
 	}
 	gotoPrev (event) {
-		if (this.props.currentImage === 0) return;
+		const { currentImage } = this.props;
+		const { imageLoaded } = this.state;
+
+		if (!imageLoaded || currentImage === 0) return;
+
 		if (event) {
 			event.preventDefault();
 			event.stopPropagation();
 		}
+
 		this.props.onClickPrev();
 		this.setState({ rotate: 0 });
-
+	}
+	closeBackdrop (event) {
+    // make sure event only happens if they click the backdrop
+    // and if the caption is widening the figure element let that respond too
+		if (event.target.id === 'lightboxBackdrop' || event.target.tagName === 'FIGURE') {
+			this.props.onClose();
+		}
 	}
 	handleKeyboardInput (event) {
 		if (event.keyCode === 37) { // left
@@ -144,20 +180,19 @@ class Lightbox extends Component {
 
 	}
 	rotate (event) {
-
 		event.preventDefault();
 		event.stopPropagation();
+
 		if (this.state.rotate === 360) {
 			this.setState({ rotate: 90 });
 		} else {
 			this.setState({ rotate: this.state.rotate + 90 });
 		}
-
 	}
 	zoom (event) {
-
 		event.preventDefault();
 		event.stopPropagation();
+
 		if (this.state.isZoomed) {
 			this.setState({
 				isZoomed: false,
@@ -169,13 +204,14 @@ class Lightbox extends Component {
 				margin: 0,
 			});
 		}
-
 	}
 	eventPreventDefault (event) {
-
 		event.preventDefault();
 		event.stopPropagation();
+	}
 
+	handleImageLoaded () {
+		this.setState({ imageLoaded: true });
 	}
 
 	// ==============================
@@ -211,44 +247,38 @@ class Lightbox extends Component {
 	renderDialog () {
 		const {
 			backdropClosesModal,
-			customControls,
 			isOpen,
-			onClose,
-			showCloseButton,
-			showRotateButton,
 			showThumbnails,
-			rotateButtonTitle,
 			width,
 		} = this.props;
+
+		const { imageLoaded } = this.state;
 
 		if (!isOpen) return <span key="closed" />;
 
 		let offsetThumbnails = 0;
 		if (showThumbnails) {
-			offsetThumbnails = theme.thumbnail.size + theme.container.gutter.vertical;
+			offsetThumbnails = this.theme.thumbnail.size + this.theme.container.gutter.vertical;
 		}
 
 		return (
 			<Container
 				key="open"
-				onClick={!!backdropClosesModal && onClose}
-				onTouchEnd={!!backdropClosesModal && onClose}
+				onClick={backdropClosesModal && this.closeBackdrop}
+				onTouchEnd={backdropClosesModal && this.closeBackdrop}
 			>
-				<div className={css(classes.content)} style={{ marginBottom: offsetThumbnails, maxWidth: width }}>
-					<Header
-						customControls={customControls}
-						onClose={onClose}
-						onRotate={this.rotate}
-						showCloseButton={showCloseButton}
-						closeButtonTitle={this.props.closeButtonTitle}
-						rotateButtonTitle={rotateButtonTitle}
-						showRotateButton={showRotateButton}
-					/>
-					{this.renderImages()}
-				</div>
-				{this.renderThumbnails()}
-				{this.renderArrowPrev()}
-				{this.renderArrowNext()}
+                <div>
+                    <div className={css(this.classes.content)} style={{ marginBottom: offsetThumbnails, maxWidth: width }}>
+                        {imageLoaded && this.renderHeader()}
+                        {this.renderImages()}
+                        {this.renderSpinner()}
+                        {imageLoaded && this.renderFooter()}
+                    </div>
+                    {imageLoaded && this.renderThumbnails()}
+                    {imageLoaded && this.renderArrowPrev()}
+                    {imageLoaded && this.renderArrowNext()}
+                    <ScrollLock />
+                </div>
 			</Container>
 		);
 	}
@@ -256,60 +286,62 @@ class Lightbox extends Component {
 		const {
 			currentImage,
 			images,
-			imageCountSeparator,
 			onClickImage,
-			showImageCount,
 			showThumbnails,
 		} = this.props;
+
+		const { imageLoaded } = this.state;
 
 		if (!images || !images.length) return null;
 
 		const image = images[currentImage];
+		image.srcSet = image.srcSet || image.srcset;
 
-		let srcset;
+		let srcSet;
 		let sizes;
 
-		if (image.srcset) {
-			srcset = image.srcset.join();
+		if (image.srcSet) {
+			srcSet = image.srcSet.join();
 			sizes = '100vw';
 		}
 
-		const thumbnailsSize = showThumbnails ? theme.thumbnail.size : 0;
-		const heightOffset = `${theme.header.height + theme.footer.height + thumbnailsSize + (theme.container.gutter.vertical)}px`;
+		const thumbnailsSize = showThumbnails ? this.theme.thumbnail.size : 0;
+		const heightOffset = `${this.theme.header.height + this.theme.footer.height + thumbnailsSize
+			+ (this.theme.container.gutter.vertical)}px`;
 
 		return (
-			<figure className={css(classes.figure)}>
+			<figure className={css(this.classes.figure)}>
 				{/*
 					Re-implement when react warning "unknown props"
 					https://fb.me/react-unknown-prop is resolved
 					<Swipeable onSwipedLeft={this.gotoNext} onSwipedRight={this.gotoPrev} />
 				*/}
-				<div className={css(classes.imageWrapper)}>
+				<div className={css(this.classes.imageWrapper)}>
 					{
 						this.props.customMedia ?
 							this.props.customMedia
-						: 
-						<img
-							className={css(classes.image)}
+						: <img
+                            className={css(this.classes.image, imageLoaded && this.classes.imageLoaded)}
 							onClick={onClickImage ? onClickImage : this.props.zoom ? this.zoom : null}
 							alt={image.alt}
 							src={image.src}
-							srcSet={srcset}
+							srcSet={srcSet}
 							style={{
 								cursor: this.props.zoom ? !this.state.isZoomed ? 'zoom-in' : 'zoom-out' : onClickImage ? 'pointer' : 'auto',
 								maxHeight: !this.state.isZoomed ? '630px' : '120vh',
 								maxWidth: !this.state.isZoomed ? '574px' : '120vh',
 								transform: !this.state.isZoomed ? `scale(1) rotate(${this.state.rotate}deg)` : `scale(1.4) rotate(${this.state.rotate}deg)`,
 								margin: this.state.margin,
-								transition: 'all .1s',
+								transition: 'all .3s',
 								display: 'inline-block',
 							}}
 						/>
+
 					}
 					{
-						this.props.content ? 
+						this.props.content ?
 							<figcaption
-								className={css(classes.figcaption)}
+								className={css(this.classes.figcaption)}
 								style={{
 									display: !this.state.isZoomed ? 'inline-block' : 'none',
 								}}
@@ -317,16 +349,9 @@ class Lightbox extends Component {
 							>
 								{this.props.children}
 							</figcaption>
-						: ''
+						: null
 					}
 				</div>
-				<Footer
-					caption={images[currentImage].caption}
-					countCurrent={currentImage + 1}
-					countSeparator={imageCountSeparator}
-					countTotal={images.length}
-					showCount={showImageCount}
-				/>
 			</figure>
 		);
 	}
@@ -344,6 +369,67 @@ class Lightbox extends Component {
 			/>
 		);
 	}
+	renderHeader () {
+		const {
+			closeButtonTitle,
+			customControls,
+			onClose,
+			showCloseButton,
+            showRotateButton,
+            rotateButtonTitle
+		} = this.props;
+
+		return (
+			<Header
+				customControls={customControls}
+				onClose={onClose}
+				showCloseButton={showCloseButton}
+				closeButtonTitle={closeButtonTitle}
+                onRotate={this.rotate}
+                rotateButtonTitle={rotateButtonTitle}
+                showRotateButton={showRotateButton}
+			/>
+		);
+	}
+	renderFooter () {
+		const {
+			currentImage,
+			images,
+			imageCountSeparator,
+			showImageCount,
+		} = this.props;
+
+		if (!images || !images.length) return null;
+
+		return (
+			<Footer
+				caption={images[currentImage].caption}
+				countCurrent={currentImage + 1}
+				countSeparator={imageCountSeparator}
+				countTotal={images.length}
+				showCount={showImageCount}
+			/>
+		);
+	}
+	renderSpinner () {
+		const {
+			spinner,
+			spinnerColor,
+			spinnerSize,
+		} = this.props;
+
+		const { imageLoaded } = this.state;
+		const Spinner = spinner;
+
+		return (
+			<div className={css(this.classes.spinner, !imageLoaded && this.classes.spinnerActive)}>
+				<Spinner
+					color={spinnerColor}
+					size={spinnerSize}
+				/>
+			</div>
+		);
+	}
 	render () {
 		return (
 			<Portal>
@@ -352,6 +438,10 @@ class Lightbox extends Component {
 		);
 	}
 }
+
+const DefaultSpinner = (props) => (
+	<ReactSpinners.BounceLoader {...props} />
+);
 
 Lightbox.propTypes = {
 	backdropClosesModal: PropTypes.bool,
@@ -363,7 +453,7 @@ Lightbox.propTypes = {
 	images: PropTypes.arrayOf(
 		PropTypes.shape({
 			src: PropTypes.string.isRequired,
-			srcset: PropTypes.array,
+			srcSet: PropTypes.array,
 			caption: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
 			thumbnail: PropTypes.string,
 		})
@@ -379,6 +469,9 @@ Lightbox.propTypes = {
 	showCloseButton: PropTypes.bool,
 	showImageCount: PropTypes.bool,
 	showThumbnails: PropTypes.bool,
+	spinner: PropTypes.func,
+	spinnerColor: PropTypes.string,
+	spinnerSize: PropTypes.number,
 	theme: PropTypes.object,
 	thumbnailOffset: PropTypes.number,
 	width: PropTypes.number,
@@ -396,6 +489,9 @@ Lightbox.defaultProps = {
 	showCloseButton: true,
 	showImageCount: true,
 	showRotateButton: true,
+	spinner: DefaultSpinner,
+	spinnerColor: 'white',
+	spinnerSize: 100,
 	theme: {},
 	thumbnailOffset: 2,
 	width: 1024,
@@ -406,7 +502,7 @@ Lightbox.childContextTypes = {
 	theme: PropTypes.object.isRequired,
 };
 
-const classes = StyleSheet.create({
+const defaultStyles = {
 	content: {
 		position: 'relative',
 	},
@@ -422,12 +518,20 @@ const classes = StyleSheet.create({
 	image: {
 		display: 'block', // removes browser default gutter
 		height: 'auto',
-		margin: '0 auto', // maintain center on very short screens OR very narrow image
+		marginLeft: 'auto', // maintain center on very short screens OR very narrow image
+		marginRight: 'auto', // maintain center on very short screens OR very narrow image
 		maxWidth: '100%',
 
 		// disable user select
 		WebkitTouchCallout: 'none',
 		userSelect: 'none',
+
+		// opacity animation on image load
+		opacity: 0,
+		transition: 'opacity 0.3s',
+	},
+	imageLoaded: {
+		opacity: 1,
 	},
 	figcaption: {
 		background: '#fff',
@@ -436,6 +540,19 @@ const classes = StyleSheet.create({
 		height: '630px',
 		overflow: 'auto',
 	},
-});
+	spinner: {
+		position: 'absolute',
+		top: '50%',
+		left: '50%',
+		transform: 'translate(-50%, -50%)',
+
+		// opacity animation to make spinner appear with delay
+		opacity: 0,
+		transition: 'opacity 0.3s',
+	},
+	spinnerActive: {
+		opacity: 1,
+	},
+};
 
 export default Lightbox;