+++ src/icons/index.js
@@ -1,6 +1,7 @@
-module.exports = {
-	arrowLeft: require('./arrowLeft'),
-	arrowRight: require('./arrowRight'),
-	close: require('./close'),
-	rotate: require('./rotate'),
+const icons = {
+    arrowLeft: require('./arrowLeft'),
+    arrowRight: require('./arrowRight'),
+    close: require('./close'),
+    rotate: require('./rotate'),
 };
+export default icons;